# doc/doc_swagger.yaml
openapi: 3.0.1
info:
  title: Cmdb API
  description: This is a simple CMDB API
  contact:
    email: guidasilvaesteves@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - description: Localhost server for testing API
    url: http://localhost:8080/api

tags:
- name: movies
- name: users

paths:
  /users:
    put:
      tags:
      - users
      summary: adds a user
      description: Adds a user to the system
      operationId: addUser
      requestBody:
        description: User to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
        required: false
      responses:
        201:
          description: user created
          content:
            application/json:
              schema:
                required:
                - token
                type: object
                properties:
                  token:
                    type: string
                    format: uuid
        400:
          description: Invalid input, object invalid
          content: {}
      x-codegen-request-body-name: user

  /movies:
    get:
      tags:
      - Movies
      summary: get movies
      description: By passing in the appropriate options, you can search for available movies
      operationId: getMovies
      security:
        - bearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
          format: uuid
      - name: searchString
        in: query
        description: pass an optional search string for looking up Movies with that text
        schema:
          type: string
      - name: skip
        in: query
        description: number of records to skip for pagination
        schema:
          minimum: 0
          type: integer
      - name: limit
        in: query
        description: maximum number of records to return
        schema:
          maximum: 50 
          minimum: 0
          type: integer
      responses:
        200:
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
        400:
          description: bad input parameter
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
    post:
      tags:
      - Movies
      summary: adds a movie
      description: Adds a movie to the system
      operationId: addMovie
      security:
        - bearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        description: movie to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMovie'
        required: false
      responses:
        201:
          description: Movie created
          content: {} 
        400:
          description: Invalid input, object invalid
          content: {}
  /tasks/{movieTitle}:
    get:
      tags:
      - Movies
      summary: Get a movie given its title
      description: Get a movie given its title
      operationId: getMoviesByName
      security:
        - bearerAuth: []
      parameters:
      - name: Authorization
        in: header
        schema:
          type: string
          format: uuid
      - name: movieTitle
        in: path
        description: ID of task that to be fetched
        required: true
        schema:
          type: integer
          minimum: 0
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        400:
          description: Invalid Title supplied
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/InvalidRequest'
                example: Invalid ID supplied
        404:
          description: Movie not found
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/InvalidRequest'
                example: Movie Not found
    put:
      tags:
      - Movies
      summary: updates a movie
      description: Updates a movie in the system
      operationId: updateGroup
      security:
        - bearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
          format: uuid
      - name: groupId
        in: path
        description: title of the movie to be deleted
        required: true
        schema:
          type: integer
      requestBody:
        description: movie to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMovie'
        required: false
      responses:
        201:
          description: movie updated
          content: {}
        404:
          description: movie not found
          content: {}
    delete:
      tags:
      - Tasks
      summary: Delete a movie by title
      description: Delete a movie by title
      operationId: deleteGroup
      security:
        - bearerAuth: []
      parameters:
      - name: movieTitle
        in: path
        description: title of the movie to be deleted
        required: true
        schema:
          type: integer
      responses:
        400:
          description: Invalid title supplied
          content: {}
        404:
          description: movie not found
          content: {}
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
  schemas:
    Movie:
      required:
      - id
      - title
      - userId
      
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Study HTTP
        description:
          type: string
          example: Task description
        userId:
          type: integer
    NewMovie:
      required:
      - title
      - description
      - userId
      type: object
      properties:
        title:
          type: string
          example: Study Http
        description:
          type: string
          example: Task description
        userId:
          type: integer
    NewUser:
      required:
      - userName
      type: object
      properties:
        userName:
          type: string
          example: joao
    InvalidRequest:
      type: object
      properties:
        error:
          type: string
          example: "Missing required parameter"
    NewMovieCreated:
      required:
      - movieid
      - status
      - movie
      type: object
      properties:
        status:
          type: string
          example: movie with title hello world created with success
        task:
          $ref: '#/components/schemas/Movie'